improve printer

+++ src/main/java/src/com/pdvsystem/Product.java
@@ -2,13 +2,13 @@ package src.com.pdvsystem;
 
 public class Product {
     private final String name;
-    private final int id;
+    private final long id;
     private final float price;
     private final String unity;
 
     private float unitiesInOrder;
 
-    public Product(int id, String name, float price, String unity){
+    public Product(long id, String name, float price, String unity){
         this.id = id;
         this.name = name;
         this.price = price;
@@ -21,7 +21,7 @@ public class Product {
         return this.name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
