+++ src/main/java/src/com/pdvsystem/ProductRepository.java
@@ -1,5 +1,5 @@
 package src.com.pdvsystem;
 
 public interface ProductRepository {
-    Product getProductById(int id);
+    Product getProductById(long id);
 }
